requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-heroku==0.3.1
django-jsonfield-backport==1.0.5
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.11.0
whitenoise==6.6.0